🚚 zb: Move TCP socket type impls to their own module
