+++ zbus/src/connection/socket/mod.rs
@@ -1,17 +1,19 @@
 mod split;
 pub use split::Split;
 
+mod tcp;
+
 #[cfg(not(feature = "tokio"))]
 use async_io::Async;
 use std::io;
+#[cfg(not(feature = "tokio"))]
+use std::sync::Arc;
 #[cfg(unix)]
 use std::{
     future::poll_fn,
     io::{IoSlice, IoSliceMut},
     task::Poll,
 };
-#[cfg(not(feature = "tokio"))]
-use std::{net::TcpStream, sync::Arc};
 
 #[cfg(all(windows, not(feature = "tokio")))]
 use uds_windows::UnixStream;
@@ -526,149 +528,6 @@ impl WriteHalf for Arc<Async<UnixStream>> {
 
     async fn peer_credentials(&mut self) -> io::Result<ConnectionCredentials> {
         ReadHalf::peer_credentials(self).await
-    }
-}
-
-#[cfg(not(feature = "tokio"))]
-#[async_trait::async_trait]
-impl ReadHalf for Arc<Async<TcpStream>> {
-    async fn recvmsg(&mut self, buf: &mut [u8]) -> RecvmsgResult {
-        match futures_util::AsyncReadExt::read(&mut self.as_ref(), buf).await {
-            Err(e) => Err(e),
-            Ok(len) => {
-                #[cfg(unix)]
-                let ret = (len, vec![]);
-                #[cfg(not(unix))]
-                let ret = len;
-                Ok(ret)
-            }
-        }
-    }
-
-    async fn peer_credentials(&mut self) -> io::Result<ConnectionCredentials> {
-        #[cfg(windows)]
-        let creds = {
-            let stream = self.clone();
-            crate::Task::spawn_blocking(
-                move || {
-                    use crate::win32::{tcp_stream_get_peer_pid, ProcessToken};
-
-                    let pid = tcp_stream_get_peer_pid(&stream.get_ref())? as _;
-                    let sid = ProcessToken::open(if pid != 0 { Some(pid as _) } else { None })
-                        .and_then(|process_token| process_token.sid())?;
-                    io::Result::Ok(
-                        ConnectionCredentials::default()
-                            .set_process_id(pid)
-                            .set_windows_sid(sid),
-                    )
-                },
-                "peer credentials",
-            )
-            .await
-        }?;
-
-        #[cfg(not(windows))]
-        let creds = ConnectionCredentials::default();
-
-        Ok(creds)
-    }
-}
-
-#[cfg(not(feature = "tokio"))]
-#[async_trait::async_trait]
-impl WriteHalf for Arc<Async<TcpStream>> {
-    async fn sendmsg(&mut self, buf: &[u8], #[cfg(unix)] fds: &[RawFd]) -> io::Result<usize> {
-        #[cfg(unix)]
-        if !fds.is_empty() {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "fds cannot be sent with a tcp stream",
-            ));
-        }
-
-        futures_util::AsyncWriteExt::write(&mut self.as_ref(), buf).await
-    }
-
-    async fn close(&mut self) -> io::Result<()> {
-        let stream = self.clone();
-        crate::Task::spawn_blocking(
-            move || stream.get_ref().shutdown(std::net::Shutdown::Both),
-            "close socket",
-        )
-        .await
-    }
-
-    async fn peer_credentials(&mut self) -> io::Result<ConnectionCredentials> {
-        ReadHalf::peer_credentials(self).await
-    }
-}
-
-#[cfg(feature = "tokio")]
-impl Socket for tokio::net::TcpStream {
-    type ReadHalf = tokio::net::tcp::OwnedReadHalf;
-    type WriteHalf = tokio::net::tcp::OwnedWriteHalf;
-
-    fn split(self) -> Split<Self::ReadHalf, Self::WriteHalf> {
-        let (read, write) = self.into_split();
-
-        Split { read, write }
-    }
-}
-
-#[cfg(feature = "tokio")]
-#[async_trait::async_trait]
-impl ReadHalf for tokio::net::tcp::OwnedReadHalf {
-    async fn recvmsg(&mut self, buf: &mut [u8]) -> RecvmsgResult {
-        use tokio::io::{AsyncReadExt, ReadBuf};
-
-        let mut read_buf = ReadBuf::new(buf);
-        self.read_buf(&mut read_buf).await.map(|_| {
-            let ret = read_buf.filled().len();
-            #[cfg(unix)]
-            let ret = (ret, vec![]);
-
-            ret
-        })
-    }
-
-    #[cfg(windows)]
-    fn peer_sid(&self) -> Option<String> {
-        use crate::win32::{socket_addr_get_pid, ProcessToken};
-
-        let peer_addr = match self.peer_addr() {
-            Ok(addr) => addr,
-            Err(_) => return None,
-        };
-
-        if let Ok(pid) = socket_addr_get_pid(&peer_addr) {
-            if let Ok(process_token) = ProcessToken::open(if pid != 0 { Some(pid) } else { None }) {
-                return process_token.sid().ok();
-            }
-        }
-
-        None
-    }
-}
-
-#[cfg(feature = "tokio")]
-#[async_trait::async_trait]
-impl WriteHalf for tokio::net::tcp::OwnedWriteHalf {
-    async fn sendmsg(&mut self, buf: &[u8], #[cfg(unix)] fds: &[RawFd]) -> io::Result<usize> {
-        use tokio::io::AsyncWriteExt;
-
-        #[cfg(unix)]
-        if !fds.is_empty() {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "fds cannot be sent with a tcp stream",
-            ));
-        }
-
-        self.write(buf).await
-    }
-
-    async fn close(&mut self) -> io::Result<()> {
-        tokio::io::AsyncWriteExt::shutdown(self).await
     }
 }
